fix deadlock in websocket api
The old api utilized DashMap as the concurrent HashMap for storing user
connections and websocket connections to the api. The problem with this
is that DashMap does not support holding locks across await points,
which this api MUST do to function properly. The lack of support for
holding locks across await points results in a deadlock when multiple
users attempt to connect to the websocket api, not just at the same
time, but at all. Therefore a migration to `scc::HashMap` was necessary,
as their implementation of a concurrent HashMap does support holding
locks across await points, as long the asynchronous api is used.